fix problem that model of stub server was in internal package (move to model package which is public)
